+++ src/shapes/Object/Object.ts
@@ -1,37 +1,37 @@
 // @ts-nocheck
-import { fabric } from '../../HEADER';
-import { cache } from '../cache';
-import { config } from '../config';
-import { ALIASING_LIMIT, iMatrix, VERSION } from '../constants';
-import { ObjectEvents } from '../EventTypeDefs';
-import { AnimatableObject } from '../mixins/object_animation.mixin';
-import { Point } from '../point.class';
-import { Shadow } from '../shadow.class';
+import { fabric } from '../../../HEADER';
+import { cache } from '../../cache';
+import { config } from '../../config';
+import { ALIASING_LIMIT, iMatrix, VERSION } from '../../constants';
+import { ObjectEvents } from '../../EventTypeDefs';
+import { Point } from '../../point.class';
+import { Shadow } from '../../shadow.class';
 import type {
+  TCacheCanvasDimensions,
   TClassProperties,
   TDegree,
   TFiller,
   TSize,
-  TCacheCanvasDimensions,
-} from '../typedefs';
-import { runningAnimations } from '../util/animation';
-import { clone } from '../util/lang_object';
-import { capitalize } from '../util/lang_string';
-import { capValue } from '../util/misc/capValue';
-import { createCanvasElement, toDataURL } from '../util/misc/dom';
+} from '../../typedefs';
+import { runningAnimations } from '../../util/animation';
+import { clone } from '../../util/lang_object';
+import { capitalize } from '../../util/lang_string';
+import { capValue } from '../../util/misc/capValue';
+import { createCanvasElement, toDataURL } from '../../util/misc/dom';
 import {
   invertTransform,
   qrDecompose,
   transformPoint,
-} from '../util/misc/matrix';
-import { enlivenObjectEnlivables } from '../util/misc/objectEnlive';
+} from '../../util/misc/matrix';
+import { enlivenObjectEnlivables } from '../../util/misc/objectEnlive';
 import {
   resetObjectTransform,
   saveObjectTransform,
-} from '../util/misc/objectTransforms';
-import { pick } from '../util/misc/pick';
-import { toFixed } from '../util/misc/toFixed';
-import type { Group } from './group.class';
+} from '../../util/misc/objectTransforms';
+import { pick } from '../../util/misc/pick';
+import { toFixed } from '../../util/misc/toFixed';
+import type { Group } from '../group.class';
+import { AnimatableObject } from './AnimatableObject';
 
 export type TCachedFabricObject = FabricObject &
   Required<
@@ -40,10 +40,13 @@ export type TCachedFabricObject = FabricObject &
       | 'zoomX'
       | 'zoomY'
       | '_cacheCanvas'
+      | '_cacheContext'
       | 'cacheTranslationX'
       | 'cacheTranslationY'
     >
-  >;
+  > & {
+    _cacheContext: CanvasRenderingContext2D;
+  };
 
 // temporary hack for unfinished migration
 type TCallSuper = (arg0: string, ...moreArgs: any[]) => any;
@@ -1971,25 +1974,18 @@ export class FabricObject<
    * @param {AbortSignal} [options.signal] handle aborting, see https://developer.mozilla.org/en-US/docs/Web/API/AbortController/signal
    * @returns {Promise<FabricObject>}
    */
-  static _fromObject<
-    T extends FabricObject,
-    X,
-    K extends X extends keyof T
-      ? { new (arg0: T[X], ...args: any[]): T }
-      : { new (...args: any[]): T }
-  >(
-    klass: K,
+  static _fromObject(
     object: Record<string, unknown>,
     { extraParam, ...options }: { extraParam?: X; signal?: AbortSignal } = {}
   ) {
-    return enlivenObjectEnlivables<InstanceType<K>>(
+    return enlivenObjectEnlivables<InstanceType<this>>(
       clone(object, true),
       options
     ).then((enlivedMap) => {
       // from the resulting enlived options, extract options.extraParam to arg0
       // to avoid accidental overrides later
       const { [extraParam]: arg0, ...rest } = { ...options, ...enlivedMap };
-      return extraParam ? new klass(arg0, rest) : new klass(rest);
+      return extraParam ? new this(arg0, rest) : new this(rest);
     });
   }
 
@@ -2004,7 +2000,7 @@ export class FabricObject<
     object: Record<string, unknown>,
     options?: { signal?: AbortSignal }
   ) {
-    return FabricObject._fromObject(FabricObject, object, options);
+    return this._fromObject(object, options);
   }
 }