+++ src/shapes/image.class.ts
@@ -17,7 +17,7 @@ import {
   LoadImageOptions,
 } from '../util/misc/objectEnlive';
 import { parsePreserveAspectRatioAttribute } from '../util/misc/svgParsing';
-import { FabricObject, fabricObjectDefaultValues } from './fabricObject.class';
+import { FabricObject, fabricObjectDefaultValues } from './Object/FabricObject';
 
 export type ImageSource =
   | HTMLImageElement
@@ -733,7 +733,7 @@ export class Image extends FabricObject {
       resizeFilter && enlivenObjects([resizeFilter], filterOptions),
       enlivenObjectEnlivables(object, options),
     ]).then(([el, filters = [], [resizeFilter] = [], hydratedProps = {}]) => {
-      return new Image(el, {
+      return new this(el, {
         ...object,
         src,
         crossOrigin,
@@ -752,7 +752,7 @@ export class Image extends FabricObject {
    * @returns {Promise<Image>}
    */
   static fromURL(url: string, options: LoadImageOptions = {}): Promise<Image> {
-    return loadImage(url, options).then((img) => new Image(img, options));
+    return loadImage(url, options).then((img) => new this(img, options));
   }
 
   /**
@@ -768,8 +768,8 @@ export class Image extends FabricObject {
     callback: (image: Image) => any,
     options: { signal?: AbortSignal } = {}
   ) {
-    const parsedAttributes = parseAttributes(element, Image.ATTRIBUTE_NAMES);
-    Image.fromURL(parsedAttributes['xlink:href'], {
+    const parsedAttributes = parseAttributes(element, this.ATTRIBUTE_NAMES);
+    this.fromURL(parsedAttributes['xlink:href'], {
       ...options,
       ...parsedAttributes,
     }).then(callback);