+++ src/shapes/path.class.ts
@@ -13,8 +13,9 @@ import {
   joinPath,
   makePathSimpler,
   parsePath,
+  type TPathSegmentsInfo,
 } from '../util/path';
-import { FabricObject, fabricObjectDefaultValues } from './fabricObject.class';
+import { FabricObject, fabricObjectDefaultValues } from './Object/FabricObject';
 
 export class Path extends FabricObject {
   /**
@@ -29,6 +30,8 @@ export class Path extends FabricObject {
   fromSVG?: boolean;
 
   sourcePath?: string;
+
+  segmentsInfo?: TPathSegmentsInfo[];
 
   /**
    * Constructor
@@ -350,7 +353,7 @@ export class Path extends FabricObject {
    * @returns {Promise<Path>}
    */
   static fromObject(object) {
-    return FabricObject._fromObject(Path, object, {
+    return this._fromObject(object, {
       extraParam: 'path',
     });
   }
@@ -365,9 +368,9 @@ export class Path extends FabricObject {
    * @param {Function} [callback] Options callback invoked after parsing is finished
    */
   static fromElement(element, callback, options) {
-    const parsedAttributes = parseAttributes(element, Path.ATTRIBUTE_NAMES);
+    const parsedAttributes = parseAttributes(element, this.ATTRIBUTE_NAMES);
     callback(
-      new Path(parsedAttributes.d, {
+      new this(parsedAttributes.d, {
         ...parsedAttributes,
         ...options,
         // we pass undefined to instruct the constructor to position the object using the bbox