+++ src/shapes/textbox.class.ts
@@ -2,9 +2,7 @@
 
 import { fabric } from '../../HEADER';
 import { TClassProperties } from '../typedefs';
-import { stylesFromArray } from '../util/misc/textStyles';
 import { IText } from './itext.class';
-import { FabricObject } from './object.class';
 import { textDefaultValues } from './text.class';
 
 /**
@@ -458,22 +456,6 @@ export class Textbox extends IText {
     return super.toObject(
       ['minWidth', 'splitByGrapheme'].concat(propertiesToInclude)
     );
-  }
-
-  /**
-   * Returns Textbox instance from an object representation
-   * @static
-   * @memberOf Textbox
-   * @param {Object} object Object to create an instance from
-   * @returns {Promise<Textbox>}
-   */
-  static fromObject(object: object): Promise<Textbox> {
-    const styles = stylesFromArray(object.styles, object.text);
-    //copy object to prevent mutation
-    const objCopy = Object.assign({}, object, { styles: styles });
-    return FabricObject._fromObject(Textbox, objCopy, {
-      extraParam: 'text',
-    });
   }
 }